Rename adapters folder to translators
